refactor: move EnsureFileServiceResolver to bootstrap
