+++ cmd/api/src/bootstrap/util.go
@@ -26,7 +26,8 @@ import (
 	"github.com/specterops/bloodhound/cmd/api/src/api/dbpool"
 	"github.com/specterops/bloodhound/cmd/api/src/api/tools"
 	"github.com/specterops/bloodhound/cmd/api/src/config"
-	"github.com/specterops/bloodhound/cmd/api/src/services/storage"
+	storageService "github.com/specterops/bloodhound/cmd/api/src/services/storage"
+	"github.com/specterops/bloodhound/packages/go/storage"
 	"github.com/specterops/dawgs"
 	"github.com/specterops/dawgs/drivers/neo4j"
 	"github.com/specterops/dawgs/drivers/pg"
@@ -39,7 +40,7 @@ import (
 // access to the FileServiceRetained, the FileServiceResolver is created prior to the
 // PreMigrationDaemons and the Entrypoint. This could then be passed in.
 type RuntimeDependencies struct {
-	FileServiceResolver storage.FileServiceResolver
+	FileServiceResolver storageService.FileServiceResolver
 }
 
 func ensureDirectory(path string) error {
@@ -48,7 +49,7 @@ func ensureDirectory(path string) error {
 			return err
 		}
 
-		if err := os.MkdirAll(path, 0755); err != nil {
+		if err := os.MkdirAll(path, 0o755); err != nil {
 			return fmt.Errorf("unable to create directory %s: %w", path, err)
 		}
 	}
@@ -81,6 +82,26 @@ func EnsureServerDirectories(cfg config.Configuration) error {
 
 	if err := ensureDirectory(cfg.CollectorsDirectory()); err != nil {
 		return err
+	}
+
+	return nil
+}
+
+var requiredFileServices = []storage.FileServiceName{
+	storage.FileServiceIngest,
+	storage.FileServiceRetained,
+	storage.FileServiceCollectors,
+	storage.FileServiceWork,
+}
+
+// EnsureFileServices confirms that the required file services are created in the supplied fileServiceResolver.
+func EnsureFileServices(
+	fileServiceResolver storageService.FileServiceResolver,
+) error {
+	for _, serviceName := range requiredFileServices {
+		if _, err := fileServiceResolver.Resolve(serviceName); err != nil {
+			return fmt.Errorf("failed to resolve %s file service: %w", serviceName, err)
+		}
 	}
 
 	return nil
